requirements.txt
pinax-theme-bootstrap
pymysql
django-material